Arrumando mensagem de erro

+++ app/src/main/java/com/example/doa_app/data/model/api/User.kt
@@ -2,7 +2,7 @@ package com.example.doa_app.data.model.api
 
 data class User(
     val id: String?,
-    val userId: Double,
+    val userId: Int,
     val name: String,
     val email: String
 )
